Food diary is mainly completed. Need to complete the details page. Passes, ktlint, detekt and all tests
