--- app/src/main/kotlin/routes/DairyRoutes.kt
@@ -1,2 +1,0 @@
-package diettracker.routes
-
